Support arbitrary props
Instead of a special monitor message for each supported property, the
monitor now sends all properties in a PropertyStore which provides typed
access to most properties and raw string access to arbitrary properties.

+++ src/command.rs
@@ -10,7 +10,7 @@ pub enum Command {
     DeviceVolumes(ObjectId, i32, i32, Vec<f32>),
     DeviceSetRoute(ObjectId, i32, i32),
     DeviceSetProfile(ObjectId, i32),
-    NodeCaptureStart(ObjectId, i32, bool),
+    NodeCaptureStart(ObjectId, u64, bool),
     NodeCaptureStop(ObjectId),
     MetadataSetProperty(ObjectId, u32, String, Option<String>, Option<String>),
 }
